+++ packages/af-utils-react-table/babel.config.js
@@ -1,10 +1,6 @@
 const plugins = [
     ["module-resolver", { root: ["./src"] }],
-    ["@babel/plugin-proposal-class-properties", { loose: true }],
-    [
-        "@babel/plugin-proposal-object-rest-spread",
-        { loose: true, useBuiltIns: true }
-    ],
+    "@babel/plugin-transform-runtime",
     [
         "transform-react-remove-prop-types",
         {
@@ -13,29 +9,17 @@ const plugins = [
     ]
 ];
 
-const presets = [["@babel/preset-react", { runtime: "automatic" }]];
+const presets = [
+    ["@babel/preset-react", { runtime: "automatic" }],
+    [
+        "@babel/preset-env",
+        {
+            loose: true
+        }
+    ]
+];
 
-module.exports = api => {
-    const presetEnvOptions = {
-        loose: true
-    };
-
-    if (api.env("test")) {
-        presetEnvOptions.targets = {
-            node: "current"
-        };
-    } else {
-        presetEnvOptions.modules = false;
-        plugins.push([
-            "@babel/plugin-transform-runtime",
-            { useESModules: true }
-        ]);
-    }
-
-    presets.push(["@babel/preset-env", presetEnvOptions]);
-
-    return {
-        plugins,
-        presets
-    };
+module.exports = {
+    plugins,
+    presets
 };