+++ packages/af-utils-react-virtual-headless/babel.config.js
@@ -1,36 +1,25 @@
 const plugins = [
-    [ "module-resolver", { "root": [ "./src" ] }],
-    [ "@babel/plugin-proposal-class-properties", { loose: true }],
-    [ "@babel/plugin-proposal-object-rest-spread", { loose: true, useBuiltIns: true }],
-    [ "transform-react-remove-prop-types", {
-        mode: "unsafe-wrap"
-    }]
+    ["module-resolver", { root: ["./src"] }],
+    "@babel/plugin-transform-runtime",
+    [
+        "transform-react-remove-prop-types",
+        {
+            mode: "unsafe-wrap"
+        }
+    ]
 ];
 
 const presets = [
-    [ "@babel/preset-react", { "runtime": "automatic" }]
+    ["@babel/preset-react", { runtime: "automatic" }],
+    [
+        "@babel/preset-env",
+        {
+            loose: true
+        }
+    ]
 ];
 
-module.exports = api => {
-
-    const presetEnvOptions = {
-        loose: true
-    };
-
-    if( api.env( "test" ) ){
-        presetEnvOptions.targets = {
-            node: "current"
-        };
-    }
-    else{
-        presetEnvOptions.modules = false;
-        plugins.push([ "@babel/plugin-transform-runtime", { useESModules: true }]);
-    }
-
-    presets.push([ "@babel/preset-env", presetEnvOptions ]);
-
-    return {
-        plugins,
-        presets
-    };
+module.exports = {
+    plugins,
+    presets
 };