+++ packages/af-utils-react-virtual-headless/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/react-virtual-headless
 
+## 0.0.3
+
+### Patch Changes
+
+-   Improved babel config, reduced bundle size a little, now @babel/runtime is not needed
+
 ## 0.0.2
 
 ### Patch Changes