+++ packages/af-utils-react-virtual-list/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @af-utils/react-virtual-list
 
+## 0.0.3
+
+### Patch Changes
+
+-   Improved babel config, reduced bundle size a little, now @babel/runtime is not needed
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.3
+    -   @af-utils/styled@0.0.2
+
 ## 0.0.2
 
 ### Patch Changes